requirements.txt
pandas==1.3.0
openai==0.27.0
python-decouple==3.5
Flask==2.0.1
flask-cors==3.0.10
pymongo==3.12.0
matplotlib
plotly
scipy
scikit-learn
dnspython